ref: move warning removal to separate class
